requirements.txt
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.4
Markdown==2.6.9
psycopg2==2.7.3.1
httpie
django-rest-auth
django-allauth